tc_board_tqm5200s_ub_comp_install.py: save u-boot.bin
if new u-boot.bin boots, save it. So, if we have an
u-boot which not boots, we know the last working
u-boot (and with it the last working commit).

With the last working commit id, we can detect the
commit, which breaks the board (using testcase
tc_board_tqm5200s_try_cur_ub.py)

Signed-off-by: Heiko Schocher <hs@denx.de>