tbotlib: rework read_end_state()
rework read_end_state() to wait really for end, without
timeouts!

First step, rework all testcases to work again with
this change.

Second step would be to remove the retry parameter.

Third step, work with this state, and see if it was
a good idea ... maybe witing endless for the endstate,
is bad ...

Signed-off-by: Heiko Schocher <hs@denx.de>