new testcase tc_workfd_check_tc_time.py
check if time for a special testcase is expired.
some testcases (like writting in a flash) are not good for
execute them every day, so give them a timeout. This testcase
checks, if the testcases is ready for a new run.
False means time is not expired
True means time is expired

new variable:
  tc_workfd_tbotfiles_dir:
  workdir, where tbot saves temporary files
  default:
  tc_workfd_work_dir + "/tmpfiles"

  tb.tc_workfd_check_tc_time_tcname:
  name of the testcase, which calls this testcase.
  used for creating different timefiles
  default:
  none, tbot fails if not set

  tc_workfd_check_tc_time_timeout:
  timeout in seconds
  default:
  none, tbot fails if not set

Use this new tescase for the bonnie test, and set
the necessary variables in board testcases, which
use the bonnie test.

Signed-off-by: Heiko Schocher <hs@denx.de>