lab_api, denx: rename 'denx' to 'ssh_std'
rename the denx lab_api to ssh_std as lab specific
task are done now in testcases. If a new lab uses
ssh for connecting to the lab PC, no need for creating
a new lb_api, only write for the lab specific tasks
a new TC (if not already one TC is working).

Signed-off-by: Heiko Schocher <hs@denx.de>