tc_uboot_check_kconfig.py: optimize testcase
first detect for all boards the architecture and compile
them, and save the md5 sums of teh binaries in a python
list. Then apply the patches, which should be checked,
and compile the boards in a second loop.

So it saves detecting the boards architecture a second
time, also the patches are only once applied.

Also no need to open a new connection for every board.

Signed-off-by: Heiko Schocher <hs@denx.de>