config: get rid of .cfg file and move to py files
use importlib.import_module for importing the variable settings,
and rename the files from *.cfg into .py. Also the variables
are now in the dictionary tb.config

config files are loaded in the following order
- board config
- lab config
- default config

If a value exists, it gets not overwritten by a following
declaration in a config file -> first has the highest
priority.

Signed-off-by: Heiko Schocher <hs@denx.de>