lab: power: add support for Gembird Silver Shield PM power controller
add support for the Gembird Silver Shield PM power controller.
More infos to this power controller found on:

http://sispmctl.sourceforge.net/

use testcase "tc_lab_sispmctl_get_variables.py" for setting
the serial and the index you need for the specific board.

This file is an example for a setup, you need to adapt
this to your needs.

Signed-off-by: Heiko Schocher <hs@denx.de>