config: add my "home" config
Use this as a base for your experiments!

replace the ip value with the value of your lab PC
replace the user with your username on the lab PC
(Do not forget to create password.py)

adapt the path settings!

The config uses the power on/off testcases for the
Gembird Silver Shield PM power controller

If you own one, and use kermit, you are ready with
a setup for your virtual lab!

Signed-off-by: Heiko Schocher <hs@denx.de>