Add CR LF heuristic to trigger response
After opening a serial connecting from the lab PC to a board it might
happen that a board does not respond to ^C to show a response, prompt or
login. So far this was handled by rebooting the board.

By adding a CR LF heuristic besides the existing ^C it is now often
possible to avoid having to power off/on the board. The advantage is
that more boards can be tested in setups which do not offer remote
power control.

Signed-off-by: Constantin Schuett <fixed-term.constantin.schuett@de.bosch.com>
Signed-off-by: Mark Jonas <mark.jonas@de.bosch.com>