tc_ub_boot_linux.py: fix for logging into bbb rootfs
on the bbb with original rootfs, there comes the message:

default username:password is [debian:temppwd]

Which breaks login into linux with tbot, as tbot
parses "assord" end send after this string the
password for login into rootfs.

So, first wait for "login:" after we got this string,
we wait for "assword".

Signed-off-by: Heiko Schocher <hs@denx.de>