tc_lx_check_reg_file.py: add a header before the devmem2 cmd
If you have to call devmem2 with a "header" set it through

tb.config.devmem2_pre

so on the bbb with original rootfs -> no devmem2 installed.

Now you can scp the devmem2 binary once to the rootfs and
set devmem2_pre='sudo pathtodevem" in your config file.

May we should enhance this tc to get devmem2 ...

Signed-off-by: Heiko Schocher <hs@denx.de>