tbotlib: allow to use python functions
now it is possible to define functions in testcases and
import them in other testcases. May it is cleaner to use
functions instead of calling python files. Both ways
are possible, time will show what;s better usuable.

Benefit from the "functions" way is, we could get rid of
tb.workfd, and the tb.config namespace ...

So, now it is possible, may I find time to convert
all testcases (which I prefer currently).

Signed-off-by: Heiko Schocher <hs@denx.de>