new tc: tc_board_bbb_restore_uboot.py: restore a broken spl/u-boot on the BBB
On the BBB we boot from sd card, if it is broken, we boot
from emmc and restore a known working uboot on the sdcard.

To switch between botmodes we can use the PIN P8_43
attached to GND -> boot from emmc, floating -> boot
from sd card.

Currently we switch with denx specific 'relais' command,
but it shows the principle and we can do it more generic.

Signed-off-by: Heiko Schocher <hs@denx.de>