documentation backend: make doc board independent
extract the DUTS_* infos from the file logfiles/duts_settings.txt
and replace DUTS_* names in the rst files with the content
of this variables.

tbots documentation backend writes the "DUTS_" Events
into the file: logfiles/duts_settings.txt

The replace_tbot_marker.py script replaces the "DUTS_*"
variables in the rst files with the content of this variables.

So it is possible to include for example the current tested
U-Boot version into the documentation, or the boardname
you tested.

With this approach you can make board independent documents.

Signed-off-by: Heiko Schocher <hs@denx.de>