new testcase: tc_connect_to_compilepc.py
Create a testcase for connecting to compile pc

This testcase creates a third connection handle to the lab and uses
ssh to create a connection to a compile pc.
The third connection can then be used with tb.workfd and tb.c_cpc
to outsource resource hungry tasks like compiling.

Signed-off-by: Heiko Schocher <hs@denx.de>
Signed-off-by: Mark Jonas <mark.jonas@de.bosch.com>