tbotlib: End tbot if unable to connect to lab pc
Previously, if tbot was unable to open a connection to the lab pc,
the occuring exception was catched but not handled correctly.
Only an error message was logged and tbot continued. When setting the
timeout for the channel an exception occured and tbot ended abruptly.

Now tbot prints at least a debugstring to the console to notify the user
why tbot is stopping.

Signed-off-by: Heiko Schocher <hs@denx.de>
Signed-off-by: Mark Jonas <mark.jonas@de.bosch.com>