documentation: add a fast test for tbot installation
add a fast test for testing tbot after you have installed
tbot. The machine where you have installed tbot is used
as lab PC and as board under test ... so you can only
do linux testcases, but good for a fast test and as
a base for setting up tbot to your needs.

Signed-off-by: Heiko Schocher <hs@denx.de>